refactor(education): add enum to string converter for cqrs response models
